refactor: _components 폴더 이동
